refactor(visits): streamline visit actions and context usage

+++ src/app/(logged-in)/visits/[id]/actions.ts
@@ -1,34 +1,36 @@
 'use server';
 
-import { notFound, redirect } from 'next/navigation';
+import { redirect } from 'next/navigation';
 import { auth } from '@/auth';
 import prisma from '@/lib/prisma';
 
-type GetVisitInput = {
-  visitId: number;
-  userId: number;
-};
-
-export const getVisit = async ({ visitId, userId }: GetVisitInput) =>
-  prisma.visit.findFirst({
-    where: { id: visitId, userId },
-  });
-
-type DeleteVisitInput = {
-  visitId: number;
-};
-
-export const deleteVisit = async ({ visitId }: DeleteVisitInput) => {
+export const getVisit = async (visitId: number) => {
   const session = await auth();
 
   if (!session?.user?.id) {
-    notFound();
+    throw new Error('Unauthorized');
   }
+
+  const userId = +session.user.id;
+
+  return prisma.visit.findFirst({
+    where: { id: visitId, userId },
+  });
+};
+
+export const deleteVisit = async (visitId: number) => {
+  const session = await auth();
+
+  if (!session?.user?.id) {
+    throw new Error('Unauthorized');
+  }
+
+  const userId = +session.user.id;
 
   await prisma.visit.deleteMany({
     where: {
       id: visitId,
-      userId: +session.user.id,
+      userId,
     },
   });
